+++ src/eca/messenger.clj
@@ -13,6 +13,7 @@
   (rewrite-content-received [this data])
   (tool-server-updated [this params])
   (config-updated [this params])
+  (provider-updated [this params])
   (showMessage [this msg])
   (progress [this params])
   (editor-diagnostics [this uri]))